requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
playwright>=1.40.0
google-genai>=0.2.0
requests>=2.31.0

